add TCP transport

+++ aucat/listen.c
@@ -19,6 +19,9 @@
 #include <sys/stat.h>
 #include <sys/un.h>
 
+#include <netinet/in.h>
+#include <netdb.h>
+
 #include <err.h>
 #include <errno.h>
 #include <fcntl.h>
@@ -29,6 +32,7 @@
 #include <unistd.h>
 
 #include "conf.h"
+#include "file.h"
 #include "listen.h"
 #include "sock.h"
 #ifdef COMPAT_STRLCPY
@@ -48,8 +52,8 @@ struct fileops listen_ops = {
 	listen_revents
 };
 
-struct listen *
-listen_new(struct fileops *ops, char *path)
+void
+listen_new_un(char *path)
 {
 	int sock, oldumask;
 	struct sockaddr_un sockname;
@@ -58,7 +62,7 @@ listen_new(struct fileops *ops, char *path)
 	sock = socket(AF_UNIX, SOCK_STREAM, 0);
 	if (sock < 0) {
 		perror("socket");
-		return NULL;
+		exit(1);
 	}
 	if (unlink(path) < 0 && errno != ENOENT) {
 		perror("unlink");
@@ -77,7 +81,7 @@ listen_new(struct fileops *ops, char *path)
 		perror("listen");
 		goto bad_close;
 	}
-	f = (struct listen *)file_new(ops, path, 1);
+	f = (struct listen *)file_new(&listen_ops, path, 1);
 	if (f == NULL)
 		goto bad_close;
 	f->path = strdup(path);
@@ -86,10 +90,70 @@ listen_new(struct fileops *ops, char *path)
 		exit(1);
 	}
 	f->fd = sock;
-	return f;
+	return;
  bad_close:
 	close(sock);
-	return NULL;
+	exit(1);	
+}
+
+void
+listen_new_tcp(char *addr, unsigned port)
+{
+	char serv[sizeof(unsigned) * 3 + 1];
+	struct addrinfo *ailist, *ai, aihints;
+	struct listen *f;
+	int s, error, opt = 1, n = 0;
+	
+	/* 
+	 * obtain a list of possible addresses for the host/port 
+	 */
+	memset(&aihints, 0, sizeof(struct addrinfo));
+	snprintf(serv, sizeof(serv), "%u", port);
+	aihints.ai_flags |= AI_PASSIVE;
+	aihints.ai_socktype = SOCK_STREAM;
+	aihints.ai_protocol = IPPROTO_TCP;
+	error = getaddrinfo(addr, serv, &aihints, &ailist);
+	if (error) {
+		fprintf(stderr, "getaddrinfo: %s\n", gai_strerror(error));
+		exit(1);
+	}
+
+	/* 
+	 * for each address, try create a listening socket bound on
+	 * that address
+	 */
+	for (ai = ailist; ai != NULL; ai = ai->ai_next) {
+		s = socket(ai->ai_family, ai->ai_socktype, ai->ai_protocol);
+		if (s < 0) {
+			perror("socket");
+			continue;
+		}
+		opt = 1;
+		if (setsockopt(s, SOL_SOCKET, SO_REUSEADDR, &opt, sizeof(int)) < 0) {
+			perror("setsockopt");
+			goto bad_close;
+		}
+		if (bind(s, ai->ai_addr, ai->ai_addrlen) < 0) {
+			perror("bind");
+			goto bad_close;
+		}
+		if (listen(s, 1) < 0) {
+			perror("listen");
+			goto bad_close;
+		}
+		f = (struct listen *)file_new(&listen_ops, addr, 1);
+		if (f == NULL) {
+		bad_close:
+			close(s);
+			continue;
+		}
+		f->path = NULL;
+		f->fd = s;
+		n++;
+	}
+	freeaddrinfo(ailist);
+	if (n == 0)
+		exit(1);
 }
 
 int
@@ -119,6 +183,7 @@ listen_revents(struct file *file, struct pollfd *pfd)
 		caddrlen = sizeof(caddrlen);
 		sock = accept(f->fd, &caddr, &caddrlen);
 		if (sock < 0) {
+			/* XXX: should kill the socket */
 			perror("accept");
 			return 0;
 		}
@@ -140,7 +205,21 @@ listen_close(struct file *file)
 {
 	struct listen *f = (struct listen *)file;
 
-	unlink(f->path);
-	free(f->path);
+	if (f->path != NULL) {
+		unlink(f->path);
+		free(f->path);
+	}
 	close(f->fd);
+}
+
+void
+listen_closeall(void)
+{
+	struct file *f, *fnext;
+
+	for (f = LIST_FIRST(&file_list); f != NULL; f = fnext) {
+		fnext = LIST_NEXT(f, entry);
+		if (f->ops == &listen_ops)
+			file_close(f);
+	}
 }
